Add DeletePatientCommand for patient deletion feature
This commit introduces a new command for deleting patients.

- Added `DeletePatientCommand` class to handle patient deletion.
- Updated `GetByIdAsync` method in `IGenericRepository` and `GenericRepository` to use `int` for patient IDs.
- Uncommented and updated HTTP DELETE method in `PatientController` to utilize the new command.
- Modified `PHMIS.csproj` to include a folder for JSON converters.
- Changed HTTP client base address in `Program.cs` to reflect new API port.

+++ PHMIS.Application/Repositories/Base/IGenericRepository.cs
@@ -35,7 +35,7 @@ namespace PHMIS.Application.Repositories.Base
             Expression<Func<T, object>>[]? includeProperties = null,
             params Expression<Func<T, object>>[] orderByProperties);
 
-        Task<T?> GetByIdAsync(long Id);
+        Task<T?> GetByIdAsync(int Id);
 
         public Task<List<T>> GetWhenAsync(Expression<Func<T, bool>> condition);
         public Task<List<T>> GetWhenContainsAsync<TProperty>(Expression<Func<T, TProperty>> property, List<TProperty> ids);
